Refactor config and agent
