+++ agent/default.go
@@ -1,0 +1,25 @@
+//  Copyright (c) 2020 Cisco and/or its affiliates.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at:
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+
+package agent
+
+var (
+	DefaultAgent Agent = NewAgent()
+)
+
+func Init(opts ...Option) { DefaultAgent.Init(opts...) }
+func Start() error        { return DefaultAgent.Start() }
+func Stop() error         { return DefaultAgent.Stop() }
+func Wait() error         { return DefaultAgent.Wait() }
+func Run() error          { return DefaultAgent.Run() }
